Drop public schema before applying scenario
