+++ support/db/dbtest/db.go
@@ -1,10 +1,46 @@
 package dbtest
 
 import (
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"os"
+	"testing"
+
 	"github.com/jmoiron/sqlx"
+	"github.com/lib/pq"
 	"github.com/stellar/go/support/db/sqlutils"
+	"github.com/stellar/go/support/errors"
 	"github.com/stretchr/testify/require"
 )
+
+// DB represents an ephemeral database that can be starts blank and can be used
+// to run tests against.
+type DB struct {
+	Dialect string
+	DSN     string
+	dbName  string
+	t       *testing.T
+	closer  func()
+	closed  bool
+}
+
+// randomName returns a new psuedo-random name that is sufficient for naming a
+// test database.  In the event that reading from the source of randomness
+// fails, a panic will occur.
+func randomName() string {
+	raw := make([]byte, 6)
+
+	_, err := rand.Read(raw)
+	if err != nil {
+		err = errors.Wrap(err, "read from rand failed")
+		panic(err)
+	}
+
+	enc := hex.EncodeToString(raw)
+
+	return fmt.Sprintf("test_%s", enc)
+}
 
 // Close closes and deletes the database represented by `db`
 func (db *DB) Close() {
@@ -45,4 +81,39 @@ func (db *DB) Open() *sqlx.DB {
 	require.NoError(db.t, err)
 
 	return conn
+}
+
+func execStatement(t *testing.T, pguser, query string) {
+	db, err := sqlx.Open("postgres", fmt.Sprintf("postgres://%s@localhost/?sslmode=disable", pguser))
+	require.NoError(t, err)
+	_, err = db.Exec(query)
+	require.NoError(t, err)
+	require.NoError(t, db.Close())
+}
+
+// Postgres provisions a new, blank database with a random name on the localhost
+// of the running process.  It assumes that you have postgres running on the
+// default port, have the command line postgres tools installed, and that the
+// current user has access to the server.  It panics on the event of a failure.
+func Postgres(t *testing.T) *DB {
+	var result DB
+	result.dbName = randomName()
+	result.Dialect = "postgres"
+	result.t = t
+
+	pgUser := os.Getenv("PGUSER")
+	if len(pgUser) == 0 {
+		pgUser = "postgres"
+	}
+
+	// create the db
+	execStatement(t, pgUser, "CREATE DATABASE "+pq.QuoteIdentifier(result.dbName))
+
+	result.DSN = fmt.Sprintf("postgres://%s@localhost/%s?sslmode=disable", pgUser, result.dbName)
+
+	result.closer = func() {
+		execStatement(t, pgUser, "DROP DATABASE "+pq.QuoteIdentifier(result.dbName))
+	}
+
+	return &result
 }
